requirements/production.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
Django==5.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
drf-spectacular==0.27.0
drf-spectacular-sidecar==2024.5.1
fonttools==4.50.0
frozenlist==1.4.0
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
Jinja2==3.1.3
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
mpld3==0.5.10
multidict==6.0.4
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
rpds-py==0.13.2
six==1.16.0
sqlparse==0.4.4
twilio==8.10.3
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.1.0
yarl==1.9.4
